[cleanup, extractor] Rename `extractors.py` to `_extractors.py`
This should be considered part of the next commit,
but is separated so that `git` can detect the renaming better
